103: Optimize performance http://code.google.com/p/sqljet/issues/detail?id=103
avoid delete at insert 

git-svn-id: http://svn.svnkit.com/repos/sqljet@559 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ trunk/sqljet/src/org/tmatesoft/sqljet/core/internal/table/SqlJetBtreeDataTable.java
@@ -459,8 +459,11 @@ public class SqlJetBtreeDataTable extends SqlJetBtreeTable implements ISqlJetBtr
             row[primaryKeyColumnNumber] = newRowId;
         }
         doActionWithIndexes(Action.UPDATE, newRowId, row);
-        cursor.delete();
-        cursor.insert(null, newRowId, pData, pData.remaining(), 0, true);
+        final boolean changeRowId = newRowId != currentRowId;
+        if (changeRowId) {
+            cursor.delete();
+        }
+        cursor.insert(null, newRowId, pData, pData.remaining(), 0, changeRowId);
         goToRow(newRowId);
 
     }
@@ -581,7 +584,7 @@ public class SqlJetBtreeDataTable extends SqlJetBtreeTable implements ISqlJetBtr
                     indexDef);
             final Object[] key = Action.DELETE == action ? null : getKeyForIndex(fields, indexDef);
             if (Action.UPDATE == action) {
-                if ( currentRowId == rowId && Arrays.deepEquals(currentKey, key)) {
+                if (currentRowId == rowId && Arrays.deepEquals(currentKey, key)) {
                     continue;
                 }
             }
